move bottom-tabs related components to separate directory, fix duplicated prop
