+++ src/index.tsx
@@ -62,8 +62,8 @@ export { default as useTransitionProgress } from './useTransitionProgress';
 /**
  * EXPERIMENTAL API BELOW. MIGHT CHANGE W/O ANY NOTICE
  */
-export { default as BottomTabs } from './components/BottomTabs';
-export { default as BottomTabsScreen } from './components/BottomTabsScreen';
+export { default as BottomTabs } from './components/bottom-tabs/BottomTabs';
+export { default as BottomTabsScreen } from './components/bottom-tabs/BottomTabsScreen';
 export { default as ScreenStackHost } from './components/gamma/ScreenStackHost';
 export {
   default as StackScreen,
